rename and refactor all monkey patch
