+++ ucm/integration/vllm/patch/patch_funcs/v0110/__init__.py
@@ -1,3 +1,4 @@
-from .load_failure_patch import _apply_load_failure_patches
+from .vllm_patch import _apply_vllm_patches
+from .vllm_ascend_patch import _apply_ascend_patches
 
-__all__ = ["_apply_load_failure_patches"]
+__all__ = ["_apply_vllm_patches", "_apply_ascend_patches"]
